requirements.txt
colpali-engine>=0.3.10
torch>=2.5.0,<2.7.0
transformers>=4.51.1,<4.52.0
gradio>=5.26.0
PyMuPDF>=1.25.5
Pillow>=10.0.0
numpy>=2.2.5
scipy>=1.15.2
peft>=0.14.0,<0.16.0
requests>=2.32.3